requirements.txt
PyYAML
pandas>=1.3.0
requests>=2.25.1
python-kucoin>=2.1.3
pytest-asyncio>=0.20.2
websocket-client>=1.3.2